requirements.txt
diff-match-patch==20121119
Django==1.8.19
django-el-pagination==3.0.1
django-import-export==0.4.2
django-multiselectfield==0.1.3
django-selectable==0.9.0
django-smart-selects==1.3.2
django-wpadmin==1.7.4
Pillow==2.9.0
psycopg2==2.6.1
sorl-thumbnail==12.3
tablib==0.11.2
wheel==0.24.0
